Created a new branch on GitHub and added UI.
Followed instructions to add UI and change game from best
of 5 to first up to 5 is the winner.
Attempting to have the text clear when a new game starts, but
not successful. Giving up on that, as this project is just practice
and I don't want to waste too much time solving an issue that I
will never have to worry about again.